# api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: Users Tasks Manager API
# during dev, should point to your local machine
host: localhost:3000
# basePath prefixes all resource paths 
basePath: /api
# 
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
paths:
  '/users':
    # binds a127 app logic to a route
    x-swagger-router-controller: users
    get:
      tags:
        - users
      summary: List all users
      description: Returns a list of users
      operationId: listAllUsers
      responses:
        '200':
          description: OK
          schema:
            type: array
            items:
              $ref: '#/definitions/User'
        default:
          description: Unexpected error
          schema:
            $ref: '#/definitions/ErrorResponse'
    post:
      tags:
        - user
      summary: Create user
      description: Creates a new user
      operationId: createUser
      parameters:
        - in: body
          name: user
          description: The user details to create
          required: true
          schema:
            $ref: '#/definitions/UserDetails'
      responses:
        '201':
          description: User created
          schema:
            $ref: '#/definitions/User'
        default:
          description: Unexpected error
          schema:
            $ref: '#/definitions/ErrorResponse'
  '/users/{id}':
    # binds a127 app logic to a route
    x-swagger-router-controller: users
    get:
      tags:
        - user
      summary: Get user info
      description: Returns a single user's information.
      operationId: getUserInfo
      parameters:
        - name: id
          in: path
          required: true
          type: string
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/User'
        default:
          description: Unexpected error
          schema:
            $ref: '#/definitions/ErrorResponse'
    put:
      tags:
        - user
        - users
      summary: Update user
      description: Updates user
      operationId: updateUser
      parameters:
        - name: id
          in: path
          required: true
          type: string
        - in: body
          name: user
          description: The user details to update
          required: true
          schema:
            $ref: '#/definitions/UserNames'
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/User'
        default:
          description: Unexpected error
          schema:
            $ref: '#/definitions/ErrorResponse'
  '/users/{user_id}/tasks':
    # binds a127 app logic to a route
    x-swagger-router-controller: tasks
    get:
      tags:
        - tasks
      summary: List all tasks for a user
      description: Returns a list of all tasks for the
        specified user
      operationId: listAllUserTasks
      parameters:
        - name: user_id
          in: path
          description: The user id to filter on
          required: true
          type: string
      responses:
        '200':
          description: OK
          schema:
            type: array
            items:
              $ref: '#/definitions/Task'
        default:
          description: Unexpected error
          schema:
            $ref: '#/definitions/ErrorResponse'
    post:
      tags:
        - task
      summary: Create task
      description: Create a new task
      operationId: createTask
      parameters:
        - name: user_id
          in: path
          required: true
          type: string
        - in: body
          name: task
          description: The task to create
          required: false
          schema:
            $ref: '#/definitions/TaskDetails'
      responses:
        '201':
          description: Task created
          schema:
            $ref: '#/definitions/Task'
        default:
          description: Unexpected arror
          schema:
            $ref: '#/definitions/ErrorResponse'
  '/users/{user_id}/tasks/{task_id}':
    # binds a127 app logic to a route
    x-swagger-router-controller: tasks
    get:
      tags:
        - task
      summary: Get task info
      description: Returns a single task's info
      operationId: getTaskInfo
      parameters:
        - name: user_id
          in: path
          required: true
          type: string
        - name: task_id
          in: path
          required: true
          type: string
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/Task'
        default:
          description: Unexpected error
          schema:
            $ref: '#/definitions/ErrorResponse'
    put:
      tags:
        - task
      summary: Update task
      description: Update a task
      operationId: updateTask
      parameters:
        - name: user_id
          in: path
          required: true
          type: string
        - name: task_id
          in: path
          required: true
          type: string
        - in: body
          name: task
          description: Change the task name
          required: false
          schema:
            $ref: '#/definitions/TaskName'
      responses:
        '200':
          description: Updated
          schema:
            $ref: '#/definitions/Task'
        default:
          description: Unexpected error
          schema:
            $ref: '#/definitions/ErrorResponse'
    delete:
      tags:
        - task
      summary: Delete task
      description: Remove a task from the list
      operationId: deleteTask
      parameters:
        - name: user_id
          in: path
          required: true
          type: string
        - name: task_id
          in: path
          required: true
          type: string
      responses:
        '200':
          description: Deleted
          schema:
            $ref: '#/definitions/Deleted'
        default:
          description: Unexpected error
          schema:
            $ref: '#/definitions/ErrorResponse'
# complex objects have schema definitions
definitions:
  User:
    type: object
    allOf:
      - type: object
        required:
          - id
          - username
        properties:
          id:
            type: string
          username:
            type: string
      - $ref: '#/definitions/UserNames'
  UserDetails:
    type: object
    allOf: 
      - type: object
        required:
          - username
        properties:
          username:
            type: string
      - $ref: '#/definitions/UserNames'
  UserNames:
    type: object
    required:
      - first_name
      - last_name
    properties:
      first_name:
        type: string
      last_name:
        type: string
  Task:
    type: object
    allOf:
      - type: object
        required:
          - id
          - user_id
          - date_time
        properties:
          id:
            type: string
          user_id:
            type: string
          status:
            type: string
            default: "pending"
            enum:
              - "pending"
              - "done"
          description:
            type: string
          date_time:
            type: string
            format: date-time
      - $ref: '#/definitions/TaskName'
  TaskDetails:
    type: object
    allOf:
      - type: object
        required:
          - description
          - date_time
        properties:
          description:
            type: string
          date_time:
            type: string
            format: pattern '^date time'
      - $ref: '#/definitions/TaskName'
  TaskName:
    type: object
    required:
      - name
    properties:
      name:
        type: string
  Deleted:
    type: object
    required:
      - collection
      - id
      - action
    properties:
      collection:
        type: string
      id:
        type: string
      action:
        type: string
        enum:
          - deleted
  HelloWorldResponse:
    required:
      - message
    properties:
      message:
        type: string
  ErrorResponse:
    required:
      - message
    properties:
      statusCode:
        type: integer
      message:
        type: string
      code:
        type: string
      stack:
        type: string